Fixes a couple of warning and add an entity to the root

+++ src/debug/ImGuiInfoWrapper.cpp
@@ -17,7 +17,7 @@ namespace ImGui {
 		const erm::Window& window = game.GetWindow();
 		
 		ImGui::SetNextWindowSize(ImVec2(600, 600), ImGuiCond_FirstUseEver);
-		ImGui::SetNextWindowPos(ImVec2(window.GetWindowWidth() / 2, window.GetWindowHeight() / 2), ImGuiCond_FirstUseEver);
+		ImGui::SetNextWindowPos(ImVec2(static_cast<float>(window.GetWindowWidth() / 2), static_cast<float>(window.GetWindowHeight() / 2)), ImGuiCond_FirstUseEver);
 		
 		if (ImGui::Begin("Info", &open, ImGuiWindowFlags_NoCollapse))
 		{
